+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -6,9 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Repurpose module to create a digitalocean droplet setup
+- Updated pre-commit hooks
 
 ## [0.0.1] - 2022-02-21
+
 ### Added
+
 - Initial module structure
 - Common examples for module usage
 - pre-commit hooks for Terraform